backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-multipart==0.0.9
openpyxl==3.1.5
PyPDF2==3.0.1
httpx==0.27.0
python-dotenv==1.0.1
google-genai>=0.2.0
pytesseract>=0.3.10
Pillow>=10.0.0
PyMuPDF>=1.23.0
reportlab>=4.0.0
